refactored: LandingPage
